refactor(Calendar): 폴더 구조 재정비 및 공통 유틸리티 분리
- CalendarFeature 폴더 구조 재정비 (Feature/, Views/, Models/)
- Reducer와 RootView 분리
- CalendarHeader, SectionComponents 컴포넌트 분리
- AppLogger를 SharedFeature로 이동 (카테고리별 로깅)
- KoreanDateFormatters를 SharedFeature로 이동 (재사용 가능)
