+++ Projects/Features/CalendarFeature/Sources/Views/WeekStripView.swift
